fix: widen the public stream parameter to intptr_t (#61)
A `cudaStream_t` is a pointer and therefore 64-bit, but every public
`stream` parameter was declared `int`. The CPU backend never dereferences
the value, so nothing here is broken at runtime -- but these declarations
are half of the hub's device-dispatch ABI: `ff::dt_euclidean` and friends in
fastfields-lib forward the same argument to either `FF_CPU::` or `FF_CUDA::`,
so the two backends must agree on its type. Widening only the CUDA side
would leave the hub unable to call both with one signature.

This is a signature-only change: 116 parameter declarations across the 8
modules, plus the `<cstdint>` includes they need. No CPU code path reads the
value, and the diff contains no non-signature lines.

Verified with `make test CXX=clang++`: 12/12 suites PASSED, 42678 checks,
0 failures -- byte-identical per-suite check counts to a pristine `main`
run of the same suite on the same machine.

Refs fastfields/fastfields-lib#4

+++ include/fastfields/api/cpu/distance.h
@@ -9,12 +9,12 @@ namespace cpu {
 void dt_euclidean(
           DLTensor & inp_out        ,
           double     voxel_spacing  = 1.0,
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 void dt_l1(
           DLTensor & inp_out        ,
           double     voxel_spacing  = 1.0,
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 void dt_spline_table(
           DLTensor & time           ,
@@ -24,7 +24,7 @@ void dt_spline_table(
     const DLTensor & times          ,
           int8_t     spline         = 3, // Cubic
           int8_t     bound          = 3, // DCT2
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 void dt_spline_brent(
           DLTensor & time           ,
@@ -36,7 +36,7 @@ void dt_spline_brent(
           double     step           ,
           int8_t     spline         = 3, // Cubic
           int8_t     bound          = 3, // DCT2
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 void dt_spline_gaussnewton(
           DLTensor & time           ,
@@ -47,7 +47,7 @@ void dt_spline_gaussnewton(
           double     tol            ,
           int8_t     spline         = 3, // Cubic
           int8_t     bound          = 3, // DCT2
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 void dt_mesh(
           DLTensor & dist           ,
@@ -57,7 +57,7 @@ void dt_mesh(
     const DLTensor & faces          ,
           bool       _signed        = true,
           bool       naive          = false,
-          int        stream         = 0
+          intptr_t   stream         = 0
 );
 
 } // namespace cpu
