+++ include/fastfields/api/cpu/reg_field.h
@@ -36,7 +36,7 @@ void field_matvec(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -52,7 +52,7 @@ void field_addmatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -68,7 +68,7 @@ void field_submatvec_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -83,7 +83,7 @@ void field_diag(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 
@@ -103,7 +103,7 @@ void field_adddiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -122,7 +122,7 @@ void field_subdiag_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -152,7 +152,7 @@ void field_kernel(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 
@@ -172,7 +172,7 @@ void field_addkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -191,7 +191,7 @@ void field_subkernel_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -219,7 +219,7 @@ void field_relax(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -243,7 +243,7 @@ void field_forward(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -266,7 +266,7 @@ void field_precond(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -285,7 +285,7 @@ void field_precond_(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -319,7 +319,7 @@ void field_matvec_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -335,7 +335,7 @@ void field_diag_rls(
     const double   * bending   = nullptr,
           int8_t     bound     = 0,
           int        ndim      = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 /**
@@ -360,7 +360,7 @@ void field_relax_rls(
           int8_t     bound     = 0,
           int        ndim      = 1,
           int        nb_iter   = 1,
-          int        stream    = 0
+          intptr_t   stream    = 0
 );
 
 } // namespace cpu
